Restructured project to only have an android version
* Moved all code from core module to android module and
removed core and desktop modules
* Replaced AbstractGPSInterface and DesktopGPSEmulator
with just AndroidGPSInterface. Renamed it GPSInterface.
