+++ android/AndroidManifest.xml
@@ -15,7 +15,7 @@
         android:label="@string/app_name"
         android:theme="@style/GdxTheme" >
         <activity
-            android:name="com.lenwohl.dreammachine.AndroidLauncher"
+            android:name=".main.AndroidLauncher"
             android:label="@string/app_name" 
             android:screenOrientation="portrait"
             android:configChanges="keyboard|keyboardHidden|orientation|screenSize">
