+++ android/src/com/lenwohl/dreammachine/gui/GUIComponent.java
