+++ android/src/com/lenwohl/dreammachine/gui/NinePatch.java
