+++ android/src/com/lenwohl/dreammachine/gui/Window.java
