+++ android/src/com/lenwohl/dreammachine/idlegame/Collector.java
