+++ android/src/com/lenwohl/dreammachine/main/AudioManager.java
