+++ android/src/com/lenwohl/dreammachine/main/RenderingManager.java
