+++ android/src/com/lenwohl/dreammachine/main/ResourceManager.java
