+++ android/src/com/lenwohl/dreammachine/scenes/Scene.java
