--- desktop/src/com/lenwohl/dreammachine/desktop/DesktopGPSEmulator.java
@@ -1,13 +1,0 @@
-package com.lenwohl.dreammachine.desktop;
-
-import com.badlogic.gdx.math.Vector2;
-import com.lenwohl.dreammachine.main.AbstractGPSInterface;
-
-public class DesktopGPSEmulator extends AbstractGPSInterface {
-
-    @Override
-	public Vector2 getCurrentGPSPosition() {
-		return virtualGPSPosition;
-	}
-	
-}
