+++ settings.gradle
@@ -1,1 +1,1 @@
-include 'desktop', 'android', 'core'
+include 'android'
